Adding routing with react-router-dom and adding auth middleware in backend.

+++ src/controllers/notes.ts
@@ -1,14 +1,14 @@
 import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
 
-    if (!notes) {
-      throw createHttpError(404, "Notes not found");
-    }
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
 
     res.status(200).json(notes);
   } catch (error) {
@@ -18,12 +18,19 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     res.status(200).json(note);
@@ -44,13 +51,17 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note must have a title.");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -77,8 +88,11 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!newTitle) {
       throw createHttpError(400, "Note must have a title.");
     }
@@ -87,6 +101,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     note.title = newTitle;
@@ -102,12 +120,19 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     const note = await NoteModel.findById(noteId).exec();
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
 
     await note.deleteOne();
